# Conflicts:
#	filter/Antinvader2015.c
#	filter/CallbackRoutine.cpp
#	filter/CallbackRoutine.h

+++ filter/ConfidentialProcess.cpp
@@ -115,12 +115,12 @@ BOOLEAN PctNewProcessDataHashNode(
                     MEM_TAG_PROCESS_TABLE);
 
     //
-    // 如果资源不足 分配失败 理论上是不大可能的
+    // 如果资源不足, 分配失败, 理论上是不大可能的.
     //
-    if (!((ULONG)ppdNewProcessData &
-         (ULONG)pwName &
-        /* (ULONG)pwPath &*/
-         (ULONG)pwMd5Digest)) {
+    if (!((SIZE_T)ppdNewProcessData &
+         (SIZE_T)pwName &
+        /* (SIZE_T)pwPath &*/
+         (SIZE_T)pwMd5Digest)) {
         return FALSE;
     }
 
@@ -252,7 +252,7 @@ ULONG  PctIsProcessDataAccordance(
     //
     // 输入的地址一定非零
     //
-	if ((!ppdProcessDataOne)||(!ppdProcessDataAnother))
+	if ((!ppdProcessDataOne) || (!ppdProcessDataAnother))
 	{
 		return CONFIDENTIAL_PROCESS_COMPARISON_NO_MATCHED;
 	}
@@ -262,7 +262,7 @@ ULONG  PctIsProcessDataAccordance(
 
     if (ulFlags & CONFIDENTIAL_PROCESS_COMPARISON_MD5) {
         //
-        // 如果需要校验Md5值, 则进行比较, 大小写不敏感，MD5值与大小写无关
+        // 如果需要校验Md5值, 则进行比较, 大小写不敏感, MD5值与大小写无关.
         //
         lRet = RtlCompareUnicodeString(&ppdProcessDataOne->usMd5Digest,         // 第一个md5
                                        &ppdProcessDataAnother->usMd5Digest,     // 第二个md5
@@ -321,11 +321,11 @@ BOOLEAN PctIsDataMachedCallback(
     )
 {
     //
-    // 使用PctIsProcessDataAccordance判断,MD5和路径暂时不启用
+    // 使用PctIsProcessDataAccordance判断, MD5和路径暂时不启用.
     //
 
-	//TO BE CONTINUE
-	//比较名称及其MD5
+	// TO BE CONTINUE
+	// 比较名称及其MD5
     return !PctIsProcessDataAccordance(
         (PCONFIDENTIAL_PROCESS_DATA)lpContext,
         (PCONFIDENTIAL_PROCESS_DATA)lpNoteData,
